+++ commands/delete.js
@@ -1,11 +1,11 @@
 const Runner = require('../src/Runner');
 const DeleteTask = require('../src/tasks/Delete');
 
-module.exports = function(args) {
+module.exports = (args) => {
 	const runner = new Runner(args);
-	const task = new DeleteTask();
-	
+
 	runner.loadConfig();
-	runner.chain([task]);
+	runner.setupLogs();
+	runner.use(new DeleteTask());
 	runner.execute();
 };