+++ commands/deploy.js
@@ -1,15 +1,17 @@
 const Runner = require('../src/Runner');
 const BuildTask = require('../src/tasks/Build');
+const ArtifactsTask = require('../src/tasks/Artifacts');
 const DeployTask = require('../src/tasks/Deploy');
 
-module.exports = function(args) {
+module.exports = (args) => {
 	const runner = new Runner(args);
 
-	const build = new BuildTask();
-	const deploy = new DeployTask();
-	
 	runner.loadConfig();
-	runner.chain([build, deploy]);
+	runner.setupLogs();
+
+	runner.use(new BuildTask());
+	runner.use(new ArtifactsTask());
+	runner.use(new DeployTask());
 
 	runner.execute();
 };