+++ commands/init.js
@@ -1,10 +1,9 @@
 const Runner = require('../src/Runner');
 const InitTask = require('../src/tasks/Init');
 
-module.exports = function(args) {
+module.exports = (args) => {
 	const runner = new Runner(args);
-	const task = new InitTask();
-	
-	runner.chain([task]);
+
+	runner.use(new InitTask());
 	runner.execute();
 };