+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v1.2.0 (2019-xx-xx)
+
+- Added `artifacts` command to upload files to a s3 bucket.
+- Added spinner to the terminal output to indicate process.
+- Added eslint config to standardise code base.
+- Reworked tasks runner to use middlewares instead of array of tasks.
+- Updated init command to use existing values if config file is already created.
+
 ## v1.1.0 (2019-01-13)
 
 - Updated **deploy** command to display stack events and wait till the update process ends.